Migrate code from runner.rb to game.rb to make runner.rb only one line, and update some tests due to changing variable/parameter names.
