+++ lib/player.rb
@@ -4,13 +4,13 @@ require 'pry'
 
 class Player
 
-    attr_reader :player_name,
+    attr_reader :name,
                 :deck
 
 
     def initialize(player_name, deck)
 
-        @player_name = player_name
+        @name = player_name
         @deck = deck
 
     end
