+++ spec/game_spec.rb
@@ -7,7 +7,7 @@ require './lib/game'
 require 'pry'
 
 
-RSpec.describe Turn do
+RSpec.describe Game do
 
     it "exists" do
 
@@ -29,40 +29,12 @@ RSpec.describe Turn do
         player1 = Player.new("Megan", deck1)
         player2 = Player.new("Aurora", deck2)
 
-        game = Game.new(player1, player2)
+        game = Game.new
 
         expect(game).to be_an_instance_of(Game)
 
     end
 
-    it "has 2 players" do
-
-        card1 = Card.new(:heart, 'Jack', 11)
-        card2 = Card.new(:heart, '10', 10)
-        card3 = Card.new(:heart, '9', 9)
-        card4 = Card.new(:diamond, 'Jack', 11)
-        card5 = Card.new(:heart, '8', 8)
-        card6 = Card.new(:diamond, 'Queen', 12)
-        card7 = Card.new(:heart, '3', 3)
-        card8 = Card.new(:diamond, '2', 2)
-
-        
-
-        deck1 = Deck.new([card1, card2, card5, card8])
-        deck2 = Deck.new([card3, card4, card6, card7])
-
-
-        player1 = Player.new("Megan", deck1)
-        player2 = Player.new("Aurora", deck2)
-
-        game = Game.new(player1, player2)
-
-
-        expect(game.player_1).to eq(player1)
-        expect(game.player_2).to eq(player2)
-
-    end
-
     it "starts game with a welcome message" do
 
         card1 = Card.new(:heart, 'Jack', 11)
@@ -83,13 +55,13 @@ RSpec.describe Turn do
         player1 = Player.new("Megan", deck1)
         player2 = Player.new("Aurora", deck2)
 
-        game = Game.new(player1, player2)
+        game = Game.new
 
         expect(game.welcome_message).to eq(nil)
 
     end
 
-    it "starts game with a welcome message" do
+    xit "starts game with a welcome message" do
 
         card1 = Card.new(:heart, 'Jack', 11)
         card2 = Card.new(:heart, '10', 10)
